feat: add denoise calib and quantize.

+++ src/model_optimizer/models/pi05/dit.py
@@ -217,7 +217,7 @@ class Pi05DenoiseStep(nn.Module, Model):
         calib_dataloader = self.get_calibrate_dataset(calib_data)
         quantize_model(self, quant_cfg, calib_dataloader)
         self.is_quantized = True
-        set_dynamic_quant(self, "fp16")
+        set_dynamic_quant(self, "bf16")
 
         self.export(export_dir, dynamo=False)
         onnx_path = f"{export_dir}/denoise.onnx"
